sm6375-common: Setup sim restart shell script
ported from sm8450
Stock MIAMI has 187 as ID, there in Lineage 185 and works great.

sm8450 and above have a very annoying SMS receiving bug where for some carriers it won't work
after a device restart if QTI telephony injection isn't available, but one of the workarounds
is to either use airplane mode toggle or eject the sim card and reinsert it

Instead of doing that? Let's try and automate disabling and enabling the sim card

Co-authored-by: Marc Bourgoin <themard85@gmail.com>
Change-Id: I76732c2a668ff179e32050aee4abe315de3d9357